requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-genai==0.1.0
PyPDF2==3.0.1
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0